Break out spawn helpers, reworks command history browsing
Removing classic js style imports and variable declarations. Added
spawn helper functions to introduce initialization code (for locating
siac binary) and simplify components code.
